feat(xod-project, xod-arduino): bus nodes
Closes #1353
