+++ packages/xod-project/src/Buses_Js.re
@@ -1,0 +1,1 @@
+let jumperizePatchRecursivelyU = Buses.jumperizePatchRecursively;
