+++ packages/xod-project/src/Node.re
@@ -2,6 +2,8 @@ type t = Js.Types.obj_val;
 
 type id = string;
 
+type label = string;
+
 type position = {
   .
   "x": int,
@@ -18,3 +20,7 @@ let create = patchPath => _create(origin, patchPath);
 [@bs.module ".."] external getId : t => id = "getNodeId";
 
 [@bs.module ".."] external getType : t => PatchPath.t = "getNodeType";
+
+[@bs.module ".."] external setType: (PatchPath.t, t) => t = "setNodeType";
+
+[@bs.module ".."] external getLabel: t => label = "getNodeLabel";
