+++ packages/xod-project/src/Node.rei
@@ -2,8 +2,14 @@ type t;
 
 type id = string;
 
+type label = string;
+
 let create: PatchPath.t => t;
 
 let getId: t => id;
 
 let getType: t => PatchPath.t;
+
+let setType: (PatchPath.t, t) => t;
+
+let getLabel: t => label;
