+++ packages/xod-project/src/Pin.rei
@@ -6,18 +6,27 @@ type direction =
   | Input
   | Output;
 
-type dataType =
+type primitiveDataType =
   | Pulse
   | Boolean
   | Number
   | Byte
   | String;
 
+/* TODO: "upgrade" to variant?
+  something like
+   | Primitive(primitiveDataType)
+   | Custom(string)
+   | Generic(genericType)
+   | Dead
+*/
+type dataType = string;
+
 let getDirection: t => direction;
 
 let normalizeLabels: list(t) => list(t);
 
-let getType: t => dataType;
+let getPrimitiveTypeExn: t => primitiveDataType;
 
 let getKey: t => key;
