+++ packages/xod-project/src/Project.re
@@ -3,15 +3,15 @@ open XodFuncTools;
 
 type t = Js.Types.obj_val;
 
-[@bs.module ".."]
-external _listPatches : t => array(Patch.t) = "listPatches";
+[@bs.module ".."] external _listPatches : t => array(Patch.t) = "listPatches";
 
 let listPatches = project => _listPatches(project) |. List.fromArray;
 
 [@bs.module ".."]
 external _listLocalPatches : t => array(Patch.t) = "listLocalPatches";
 
-let listLocalPatches = project => _listLocalPatches(project) |. List.fromArray;
+let listLocalPatches = project =>
+  _listLocalPatches(project) |. List.fromArray;
 
 [@bs.module ".."]
 external _assocPatch : (PatchPath.t, Patch.t, t) => Either.t(Js.Exn.t, t) =
@@ -29,3 +29,17 @@ let getPatchByPath = (project, path) =>
 
 let getPatchByNode = (project, node) =>
   getPatchByPath(project, Node.getType(node));
+
+[@bs.module ".."]
+external _getPatchDependencies : (Patch.path, t) => array(Patch.path) =
+  "getPatchDependencies";
+
+let getPatchDependencies = (project, patchPath) =>
+  _getPatchDependencies(patchPath, project) |. Belt.List.fromArray;
+
+[@bs.module ".."]
+external _assocPatchListUnsafe : (array(Patch.t), t) => t =
+  "assocPatchListUnsafe";
+
+let assocPatchList = (project, patchList) =>
+  _assocPatchListUnsafe(Belt.List.toArray(patchList), project);
