+++ packages/xod-project/src/Project.rei
@@ -11,3 +11,7 @@ let assocPatch: (t, PatchPath.t, Patch.t) => XResult.t(t);
 let getPatchByPath: (t, PatchPath.t) => option(Patch.t);
 
 let getPatchByNode: (t, Node.t) => option(Patch.t);
+
+let getPatchDependencies: (t, Patch.path) => list(Patch.path);
+
+let assocPatchList: (t, list(Patch.t)) => t;
