+++ packages/xod-project/src/TypeDeduction_Js.re
@@ -1,0 +1,6 @@
+let deducePinTypes:
+  (Patch.t, Project.t) => TypeDeduction.ResultsMap.t_Js =
+  (patch, project) =>
+    patch
+    |. TypeDeduction.deducePinTypes(project)
+    |. TypeDeduction.ResultsMap.toJsDict;
