+++ packages/xod-project/src/constants.js
@@ -191,6 +191,11 @@ export const OUTPUT_SELF_PATH = 'xod/patch-nodes/output-self';
  */
 export const UTILITY_MARKER_PATH = 'xod/patch-nodes/utility';
 
+export const FROM_BUS_PATH = 'xod/patch-nodes/from-bus';
+export const TO_BUS_PATH = 'xod/patch-nodes/to-bus';
+
+export const JUMPER_PATCH_PATH = 'xod/core/jumper';
+
 export const IMPL_FILENAME = 'patch.cpp';
 
 export const UNTITLED_PROJECT = 'untitled';
