+++ packages/xod-tabtest/package.json
@@ -14,6 +14,12 @@
   "author": "XOD",
   "license": "AGPL-3.0",
   "main": "src/Tabtest_Js.bs.js",
+  "dependencies": {
+    "holes": "^0.1.0",
+    "xod-func-tools": "^0.21.0",
+    "xod-project": "^0.22.0",
+    "xod-arduino": "^0.22.0"
+  },
   "devDependencies": {
     "@glennsl/bs-jest": "^0.4.2",
     "bs-platform": "^3.1.5"
