+++ workspace/__lib__/xod/core/jumper/patch.xodp
@@ -1,0 +1,41 @@
+{
+  "links": [
+    {
+      "id": "HkrhyNMEQ",
+      "input": {
+        "nodeId": "__out__",
+        "pinKey": "__in__"
+      },
+      "output": {
+        "nodeId": "__in__",
+        "pinKey": "__out__"
+      }
+    }
+  ],
+  "nodes": [
+    {
+      "id": "B1XUkEfEQ",
+      "position": {
+        "x": 170,
+        "y": 0
+      },
+      "type": "xod/patch-nodes/utility"
+    },
+    {
+      "id": "__in__",
+      "position": {
+        "x": 34,
+        "y": 0
+      },
+      "type": "xod/patch-nodes/input-t1"
+    },
+    {
+      "id": "__out__",
+      "position": {
+        "x": 34,
+        "y": 204
+      },
+      "type": "xod/patch-nodes/output-t1"
+    }
+  ]
+}
